Make Fragment a first-class IR citizen: move it into ir/kernel/ir.py
Per review: Fragment belongs with the fragment nodes it emits / is linked to —
FragmentApply / FragmentRowReduce / FragmentMask + RegFragment / LdmatrixLoad /
RegStore all live in ir/kernel/ir.py. Move Fragment (+ FragmentPhases +
_frag_weight_name) there too and delete the assembly/_frag_softmax.py module;
the fragment machinery is now in one place.

- ir/kernel/ir.py gains Fragment / FragmentPhases (next to the fragment ops +
  storage nodes); split_carrier is lazily imported in combine() to avoid any
  load-order cycle (carrier_algebra is pure ir.stmt).
- assembly/_assemble.py imports Fragment from ir.kernel.ir (alongside FragmentApply
  etc.); _frag_softmax.py deleted.
- tests/compiler/passes/test_frag_softmax.py -> test_fragment.py (imports Fragment
  from ir.kernel.ir). Docstring refs across carrier_algebra / leaves / the e2e test
  + the tile ARCHITECTURE row updated to ir/kernel/ir.Fragment.

m16n8 byte-identical; GPU flash e2e green; full compiler suite 1735 passed; lint clean.
