+++ deplodock/compiler/ir/stmt/carrier_algebra.py
@@ -15,7 +15,7 @@ how state and partials are distributed):
   fold, inline in the loop body).
 - **cross-thread** (lanes / smem) — ``lowering/kernel/_combine.emit_combine`` → ``WarpShuffle`` /
   ``TreeHalve`` (the cooperative reduce of a partial split across the CTA's threads).
-- **fragment** (m16n8 tensor-core registers) — ``assembly/_frag_softmax.realize_fragment_softmax``
+- **fragment** (m16n8 tensor-core registers) — ``ir/kernel/ir.Fragment.combine``
   → ``FragmentRowReduce`` (the cross-column fold) + the carrier-generic ``FragmentApply`` (every
   pointwise step — ``exp`` / scale / any op, the warp-chain flash softmax and beyond).
 
@@ -27,7 +27,7 @@ abstract-interprets a merge program over a :class:`Distribution` **backend**, ta
 algebra to a target distribution by the distribution law — a reduce over the distributed axis
 becomes the backend's cross-partition combine (``Distribution.fold``), and an elementwise op the
 backend's per-element map (``Distribution.pointwise``); scalars / carried-state reassigns stay
-replicated. The fragment realizer (``assembly/_frag_softmax.FragmentDist``) is one backend — its
+replicated. The fragment realizer (``ir/kernel/ir.Fragment``) is one backend — its
 ``fold`` is a ``FragmentRowReduce`` over the C-fragment columns, its ``pointwise`` a
 ``FragmentApply`` over the registers. No softmax knowledge in the driver: it dispatches by the
 op's *role under the distribution*, not by name. (The cross-thread ``emit_combine`` is a different
